fixed some integer type casting mismatches, removed a lot of the excess 'as u32', etc. and switched to (almost entirely) u32 integer types

+++ src/lib.rs
@@ -7,6 +7,6 @@ pub use algorithms::threshold;
 #[derive(Debug)]
 pub struct GrayImageBuffer {
     pub data: Vec<u8>,
-    pub width: usize,
-    pub height: usize
+    pub width: u32,
+    pub height: u32
 }
